requirements.txt
amqp==2.6.0
anyjson==0.3.3
appdirs==1.4.3
billiard==3.6.3.0
celery==4.4.5
certifi==2019.11.28
chardet==3.0.4
distlib==0.3.0
dj-database-url==0.5.0
Django==2.2.11
django-appconf==1.0.4
django-celery==3.3.1
django-celery-email==3.0.0
django-ckeditor==5.9.0
django-jalali==3.3.0
django-js-asset==1.2.2
django-mailgun==0.9.1
filelock==3.0.12
future==0.18.2
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.4.0
jdatetime==3.6.2
kombu==4.6.10
Pillow==7.0.0
pipenv==2018.11.26
python-decouple==3.3
pytz==2019.3
redis==3.5.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
vine==1.3.0
virtualenv==20.0.16
virtualenv-clone==0.5.4
zipp==1.2.0
